+++ src/bpms/BpmsEngine.ts
@@ -6,7 +6,7 @@ import { DmnEngine, DmnEngineOptions } from './dmn';
 import { HistoryService, HistoryServiceOptions } from './history/HistoryService';
 import { IdentityService, IdentityServiceOptions } from './identity/IdentityService';
 import { JobService, JobServiceOptions } from './job/JobService';
-import { NavigationService, NavigationServiceOptions } from './navigation/NavigationService';
+import { RouterService, RouterServiceOptions } from './router/RouterService';
 import { NotificationService } from './notification/NotificationService';
 import { QueryService, QueryServiceOptions } from './query/QueryService';
 import { TaskService, TaskServiceOptions } from './task/TaskService';
@@ -14,6 +14,7 @@ import { TenantService, TenantServiceOptions } from './tenant/TenantService';
 import { UIService, UIServiceOptions } from './ui/UIService';
 import { DataModelServiceOptions, DataModelService } from './data/DataModelService';
 import { DataSourceServiceOptions, DataSourceService } from './data/DataSourceService';
+import { OrganizationService, OrganizationServiceOptions } from './organization/OrganizationService';
 export interface BpmsEngineOptions {
     name: string;
     cache?: boolean;
@@ -23,10 +24,11 @@ export interface BpmsEngineOptions {
     datamodelService?: DataModelServiceOptions;
     datasourceService?: DataSourceServiceOptions;
     identityService?: IdentityServiceOptions;
+    organizationService?: OrganizationServiceOptions;
     historyService?: HistoryServiceOptions;
     taskService?: TaskServiceOptions;
     tenantService?: TenantServiceOptions;
-    navigationService?: NavigationServiceOptions;
+    routerService?: RouterServiceOptions;
     uiService?: UIServiceOptions;
     jobService?: JobServiceOptions;
     queryService?: QueryServiceOptions;
@@ -50,12 +52,13 @@ export class BpmsEngine {
     private datasourceService!: DataSourceService;
     private historyService!: HistoryService;
     private identityService!: IdentityService;
+    private organizationService!: OrganizationService;
     private taskService!: TaskService;
     private tenantService!: TenantService;
     private uiService!: UIService;
     private queryService!: QueryService;
     private jobService!: JobService;
-    private navigationService!: NavigationService;
+    private routerService!: RouterService;
     private notificationService!: NotificationService;
 
     private id: string = uuidv1();
@@ -106,6 +109,10 @@ export class BpmsEngine {
             name: this.name,
             ...this.options.identityService,
         });
+        this.organizationService = OrganizationService.createService(this, {
+            name: this.name,
+            ...this.options.organizationService,
+        });
         this.historyService = HistoryService.createService(this, {
             name: this.name,
             ...this.options.historyService,
@@ -122,9 +129,9 @@ export class BpmsEngine {
             name: this.name,
             ...this.options.jobService,
         });
-        this.navigationService = NavigationService.createService(this, {
+        this.routerService = RouterService.createService(this, {
             name: this.name,
-            ...this.options.navigationService,
+            ...this.options.routerService,
         });
         this.notificationService = NotificationService.createService(this, {
             name: this.name,
@@ -207,8 +214,8 @@ export class BpmsEngine {
         return this.datasourceService as any;
     }
 
-    public get NavigationService(): NavigationService {
-        return this.navigationService as any;
+    public get RouterService(): RouterService {
+        return this.routerService as any;
     }
     public get UIService(): UIService {
         return this.uiService as any;
@@ -231,6 +238,9 @@ export class BpmsEngine {
         return this.taskService as any;
     }
     public get IdentityService(): IdentityService {
+        return this.identityService as any;
+    }
+    public get OrganizationService(): OrganizationService {
         return this.identityService as any;
     }
     public get HistoryService(): HistoryService {