+++ src/bpms/bpmn/BpmnEngine.ts
@@ -44,11 +44,14 @@ export interface BpmsBpmnDefinition {
     name: string;
     definitions: any;
 
+    version: number;
+
     createdAt?: Date;
 }
 
 export interface BpmsProcess {
     id?: string;
+    definitionId?: string;
     name: string;
     state: string;
     stopped: boolean;
@@ -136,6 +139,15 @@ export class BpmnEngine {
         const f = await this.bpmnDefinitionRepository.find({ name });
         if (!f) {
             const r = await this.bpmnDefinitionRepository.create({ definitions: source, name });
+            this.bpmsEngine?.HistoryService.create({
+                type: 'info',
+                source: this.name,
+                message: `The process definition has been created`,
+                data: {
+                    method: 'createDefinitions',
+                },
+                eventId: 10031,
+            });
             return r;
         }
         throw new Error('The bpmn definition already exists');
@@ -215,18 +227,47 @@ export class BpmnEngine {
         const p = new Promise<BpmnProcessInstance>(async (resolve, reject) => {
             try {
                 // using  registered definition if name already registered .
-                if (options && options.name && !options.source) {
-                    const d = await this.bpmnDefinitionRepository.find({
-                        name: options.name,
-                    });
+                if (options && !options.source) {
+                    const d = options.definitionId
+                        ? await this.bpmnDefinitionRepository.find({
+                              id: options.definitionId,
+                          })
+                        : await this.bpmnDefinitionRepository.find({
+                              name: options.definitionName,
+                              version: options.definitionVersion,
+                          });
                     if (d) {
+                        options.definitionId = d.id;
+                        options.definitionName = d.name;
+                        options.definitionVersion = d.version;
                         options.source = d.definitions;
                     }
                 }
                 const proc = new BpmnProcessInstance(self, options);
                 await this.loadedProcessRepository.update(proc.Id, proc, true);
+                await this.persistProcess({ filter: { id: proc.Id } });
+                this.bpmsEngine?.HistoryService.create({
+                    type: 'info',
+                    source: this.name,
+                    message: `The process has been created`,
+                    data: {
+                        definitionId: options?.definitionId,
+                        processId: proc.Id,
+                        method: 'createProcess',
+                    },
+                    eventId: 10032,
+                });
                 proc.onEnd(async () => {
                     await this.loadedProcessRepository.delete(proc.Id);
+                    this.bpmsEngine?.HistoryService.create({
+                        type: 'info',
+                        source: this.name,
+                        message: `The process has been terminated`,
+                        data: {
+                            processId: proc.Id,
+                        },
+                        eventId: 10033,
+                    });
                 });
                 resolve(proc);
             } catch (error) {
@@ -269,6 +310,7 @@ export class BpmnEngine {
                     if (!clist.some(xx => xx.Id === pitem.id)) {
                         const p = await this.createProcess({
                             name: pitem.name,
+                            definitionId: pitem.definitionId,
                             id: pitem.id,
                         });
                         p.recover(pitem);
@@ -284,6 +326,7 @@ export class BpmnEngine {
                     if (!clist.some(xx => xx.Id === pitem.id)) {
                         const p = await this.createProcess({
                             name: pitem.name,
+                            definitionId: pitem.definitionId,
                             id: pitem.id,
                         });
                         p.recover(pitem, { listener: p });
@@ -318,6 +361,9 @@ export class BpmnEngine {
                             name: p.Name,
                             state: p.State,
                             stopped: p.Stopped,
+                            definitionId: p.DefinitionId,
+                            definitionName: p.DefinitionName,
+                            definitionVersion: p.DefinitionVersion,
                             persistedAt: new Date(),
                             data: d,
                         },
@@ -341,6 +387,9 @@ export class BpmnEngine {
                                 name: p.Name,
                                 state: p.State,
                                 stopped: p.Stopped,
+                                definitionId: p.DefinitionId,
+                                definitionName: p.DefinitionName,
+                                definitionVersion: p.DefinitionVersion,
                                 persistedAt: new Date(),
                                 data: d,
                             },