+++ src/bpms/history/HistoryService.ts
@@ -3,24 +3,28 @@
 import { uuidv1 } from 'nowjs-core/lib/utils';
 import { BpmsEngine } from '../BpmsEngine';
 import { TaskRepository } from '../task/TaskRepository';
-import { HistoryData, HistoryMemoryRepository, HistoryRepository } from './HistoryRepository';
+import { BpmsHistoryModel, HistoryMemoryRepository, HistoryRepository } from './HistoryRepository';
 import { QueryOptions, QueryResult, ScalarOptions, IdExpression, FilterExpression } from '../data/Repository';
+import { BpmsService } from '../BpmsService';
 
 export interface HistoryServiceOptions {
     historyRepository?: HistoryRepository;
     name: string;
 }
 
-export interface BpmsHistoryEntity {
+export interface BpmsHistoryEntry {
     id?: string;
     createdAt?: string;
-    source?: string;
-    tenantId: string;
-    userId: string;
+    source: string;
+    eventId: number;
+    message: string;
+    data?: Record<string,any>;
+    tenantId?: string;
+    userId?: string;
 
-    type?: string;
+    type: 'info'|'warn'|'error'|'fatal';
 }
-export class HistoryService<T extends BpmsHistoryEntity = BpmsHistoryEntity> {
+export class HistoryService<T extends BpmsHistoryEntry = BpmsHistoryEntry> implements BpmsService {
     private historyRepository: HistoryRepository<T>;
     private id: string = uuidv1();
     private options: HistoryServiceOptions;